Silence warnings

+++ src/core/vsapi.cpp
@@ -277,7 +277,7 @@ static char VS_CC propGetType(const VSMap *map, const char *key) {
     try { \
         VSVariant &l = map->at(key); \
         if (l.getType() == (vt)) { \
-            if (index >= 0 && index < l.size()) { \
+            if (index >= 0 && static_cast<size_t>(index) < l.size()) { \
                 if (error) \
                     *error = 0; \
                 return (retexpr); \
