+++ src/core/vscore.cpp
@@ -1199,7 +1199,7 @@ void VSPlugin::registerFunction(const std::string &name, const std::string &args
 static bool hasCompatNodes(const VSMap &m) {
     for (const auto &vsv : m.getStorage()) {
         if (vsv.second.getType() == VSVariant::vNode) {
-            for (int i = 0; i < vsv.second.size(); i++) {
+            for (size_t i = 0; i < vsv.second.size(); i++) {
                 for (size_t j = 0; j < vsv.second.getValue<VSNodeRef>(i).clip->getNumOutputs(); j++) {
                     const VSNodeRef &ref = vsv.second.getValue<VSNodeRef>(i);
                     const VSVideoInfo &vi = ref.clip->getVideoInfo(static_cast<int>(j));
@@ -1215,7 +1215,7 @@ static bool hasCompatNodes(const VSMap &m) {
 static bool hasForeignNodes(const VSMap &m, const VSCore *core) {
     for (const auto &vsv : m.getStorage()) {
         if (vsv.second.getType() == VSVariant::vNode) {
-            for (int i = 0; i < vsv.second.size(); i++) {
+            for (size_t i = 0; i < vsv.second.size(); i++) {
                 for (size_t j = 0; j < vsv.second.getValue<VSNodeRef>(i).clip->getNumOutputs(); j++) {
                     const VSNodeRef &ref = vsv.second.getValue<VSNodeRef>(i);
                     if (!ref.clip->isRightCore(core))
